+++ src/primaite/simulator/network/hardware/network_interface/wireless/wireless_nic.py
@@ -80,4 +80,4 @@ class WirelessNIC(IPWirelessNetworkInterface):
 
         :return: A string combining the port number, MAC address and IP address of the NIC.
         """
-        return f"Port {self.port_num}: {self.mac_address}/{self.ip_address}"
+        return f"Port {self.port_name if self.port_name else self.port_num}: {self.mac_address}/{self.ip_address}"